chore: move theme switch to full stack component
